[BF] minor fixes with ask world

+++ src/main/kotlin/space/yaroslav/familybot/common/Models.kt
@@ -37,7 +37,7 @@ data class AskWorldQuestion(
         val user: User,
         val chat: Chat,
         val date: Instant,
-        val messageId: Int?
+        val messageId: Long?
 )
 
 data class AskWorldReply(
